+++ packages/tracker-core/src/announce_handler.rs
@@ -99,8 +99,8 @@ use torrust_tracker_primitives::core::AnnounceData;
 use torrust_tracker_primitives::peer;
 
 use super::torrent::repository::in_memory::InMemoryTorrentRepository;
-use super::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
 use crate::error::AnnounceError;
+use crate::statistics::persisted::downloads::DatabaseDownloadsMetricRepository;
 use crate::whitelist::authorization::WhitelistAuthorization;
 
 /// Handles `announce` requests from `BitTorrent` clients.
