+++ packages/tracker-core/src/container.rs
@@ -11,9 +11,9 @@ use crate::authentication::service::AuthenticationService;
 use crate::databases::setup::initialize_database;
 use crate::databases::Database;
 use crate::scrape_handler::ScrapeHandler;
+use crate::statistics::persisted::downloads::DatabaseDownloadsMetricRepository;
 use crate::torrent::manager::TorrentsManager;
 use crate::torrent::repository::in_memory::InMemoryTorrentRepository;
-use crate::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
 use crate::whitelist::authorization::WhitelistAuthorization;
 use crate::whitelist::manager::WhitelistManager;
 use crate::whitelist::repository::in_memory::InMemoryWhitelist;
