+++ packages/tracker-core/src/statistics/persisted/downloads.rs
