+++ packages/tracker-core/src/statistics/persisted/mod.rs
@@ -1,3 +1,5 @@
+pub mod downloads;
+
 use std::sync::Arc;
 
 use thiserror::Error;
@@ -8,7 +10,7 @@ use torrust_tracker_primitives::DurationSinceUnixEpoch;
 use super::repository::Repository;
 use super::TRACKER_CORE_PERSISTENT_TORRENTS_DOWNLOADS_TOTAL;
 use crate::databases;
-use crate::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
+use crate::statistics::persisted::downloads::DatabaseDownloadsMetricRepository;
 
 /// Loads persisted metrics from the database and sets them in the stats repository.
 ///
