+++ packages/tracker-core/src/torrent/repository/mod.rs
@@ -1,3 +1,2 @@
 //! Torrent repository implementations.
 pub mod in_memory;
-pub mod downloads;
