+++ packages/udp-tracker-server/src/handlers/announce.rs
@@ -835,8 +835,8 @@ mod tests {
                 use aquatic_udp_protocol::{InfoHash as AquaticInfoHash, PeerId as AquaticPeerId};
                 use bittorrent_tracker_core::announce_handler::AnnounceHandler;
                 use bittorrent_tracker_core::databases::setup::initialize_database;
+                use bittorrent_tracker_core::statistics::persisted::downloads::DatabaseDownloadsMetricRepository;
                 use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
-                use bittorrent_tracker_core::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
                 use bittorrent_tracker_core::whitelist::authorization::WhitelistAuthorization;
                 use bittorrent_tracker_core::whitelist::repository::in_memory::InMemoryWhitelist;
                 use bittorrent_udp_tracker_core::connection_cookie::{gen_remote_fingerprint, make};
