+++ packages/udp-tracker-server/src/handlers/mod.rs
@@ -211,8 +211,8 @@ pub(crate) mod tests {
     use bittorrent_tracker_core::announce_handler::AnnounceHandler;
     use bittorrent_tracker_core::databases::setup::initialize_database;
     use bittorrent_tracker_core::scrape_handler::ScrapeHandler;
+    use bittorrent_tracker_core::statistics::persisted::downloads::DatabaseDownloadsMetricRepository;
     use bittorrent_tracker_core::torrent::repository::in_memory::InMemoryTorrentRepository;
-    use bittorrent_tracker_core::torrent::repository::downloads::DatabaseDownloadsMetricRepository;
     use bittorrent_tracker_core::whitelist;
     use bittorrent_tracker_core::whitelist::authorization::WhitelistAuthorization;
     use bittorrent_tracker_core::whitelist::repository::in_memory::InMemoryWhitelist;
